[OTA] Move OTARequestorInterface to clusters directory
