+++ src/platform/ExtendedOTARequestorDriver.cpp
@@ -14,8 +14,9 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
+
+#include <app/clusters/ota-requestor/OTARequestorInterface.h>
 #include <platform/ExtendedOTARequestorDriver.h>
-#include <platform/OTARequestorInterface.h>
 
 namespace chip {
 namespace DeviceLayer {
