+++ src/platform/P6/OTAImageProcessorImpl.cpp
@@ -17,7 +17,6 @@
  */
 
 #include <app/clusters/ota-requestor/OTADownloader.h>
-#include <platform/OTARequestorInterface.h>
 
 #include "OTAImageProcessorImpl.h"
